*Actually* fix zsh completions
This is much more comfortable - the `ac` command has been changed from
an alias to a function, which makes it much easier to work with. I've
also added a check to assert that each tool exists prior to generating
its completion.

Further, completions are no longer going to be tracked in this repo.
This makes sense, since the completions for open source tools are likely
to change often.